[WIP] Add first version of observed take pose state
Custom state to put the robot in position while tracking the error between observed and control robot.
This error is induced by the admittance control from the stabilizer: pushing the feet in the ground
results in a high position error on the hands transform tasks (the error is tracked using the false control instance).
This version still uses the bspline trajectory which might cause divergence (bspline computes the trajectory including time).